feat/#44-DeletePopup UI 수정 및 팝업 컴포넌트 폴더링
